[WNMGDS 968] convert button component to typescript (#1092)
* Convert Button non-component files to Typescript.

* Update snapshot to TSX.

* Refactor Button component to use TS.

* Remove type definition file.

* Remove event types from button eventhandlers.

* Adding 'onClick' to OmitProps.

* Add anchor tag to Omit, string type to children, and add button/link omitprops to omit type.

* Remove comment.
