+++ packages/design-system-docs/src/pages/components/Button/_Button.docs.scss
@@ -45,7 +45,7 @@ you could pass in a `target` prop to pass to the rendered anchor element.
 
 @react-example Button.example.jsx
 
-@react-props Button.jsx
+@react-props Button.tsx
 
 Style guide: components.button.react
 */
