+++ packages/design-system/src/components/Button/Button.tsx
@@ -1,9 +1,65 @@
-import PropTypes from 'prop-types';
 import React from 'react';
 import classNames from 'classnames';
 
-export class Button extends React.PureComponent {
-  constructor(props) {
+export interface ButtonProps {
+  /**
+   * Label text or HTML
+   */
+  children: string | React.ReactNode;
+  /**
+   * Additional classes to be added to the root button element.
+   * Useful for adding utility classes.
+   */
+  className?: string;
+  /**
+   * When provided, this will render the passed in component. This is useful when
+   * integrating with React Router's `<Link>` or using your own custom component.
+   */
+  component?: React.ReactElement<any> | any | ((...args: any[]) => any);
+  disabled?: boolean;
+  /**
+   * When provided the root component will render as an `<a>` element
+   * rather than `button`.
+   */
+  href?: string;
+  /**
+   * Access a reference to the `button` or `a` element
+   */
+  inputRef?: (...args: any[]) => any;
+  /**
+   * @hide-prop [Deprecated] Use inversed instead
+   */
+  inverse?: boolean;
+  /** Applies the inverse theme styling */
+  inversed?: boolean;
+  /**
+   * Returns the [`SyntheticEvent`](https://facebook.github.io/react/docs/events.html).
+   * Not called when the button is disabled.
+   */
+  onClick?: (...args: any[]) => any;
+  size?: 'small' | 'big';
+  /**
+   * Button [`type`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-type) attribute
+   */
+  type?: 'button' | 'submit';
+  /**
+   * A string corresponding to the button-component variation classes.
+   * The `'danger'` variation is deprecated and will be removed in a future release.
+   */
+  variation?: 'primary' | 'danger' | 'success' | 'transparent';
+}
+
+type OmitProps = 'children' | 'className' | 'onClick' | 'ref' | 'size' | 'type';
+
+export default class Button extends React.PureComponent<
+  Omit<React.ComponentPropsWithRef<'button' | 'a'>, OmitProps> & ButtonProps
+> {
+  static defaultProps = {
+    type: 'button',
+    component: 'button',
+  };
+
+  constructor(props: ButtonProps) {
     super(props);
 
     if (process.env.NODE_ENV !== 'production') {
@@ -24,13 +80,13 @@ export class Button extends React.PureComponent {
   }
 
   // Get an object of props to pass to the rendered <Button> component
-  attrs() {
+  attrs(): any {
     /**
      * Since any number of arbitrary props can be passed into this component, we
      * use a destructuring assignment to get only the props we want to pass to the
      * rendered HTML element. For example, the "variation" prop is used to generate
-     * the classNames, but doesn't need passed to the rendered component, so we
-     * omit it here so that it's not included in the props object.
+     * the classNames, but doesn't need to be passed to the rendered component, so
+     * we omit it here so that it's not included in the props object.
      */
     const {
       className,
@@ -44,7 +100,7 @@ export class Button extends React.PureComponent {
       ...props
     } = this.props;
 
-    const attrs = {
+    const attrs: any = {
       className: this.classNames(),
       ...props,
     };
@@ -63,8 +119,9 @@ export class Button extends React.PureComponent {
     return attrs;
   }
 
-  componentType() {
+  componentType(): string {
     let component = this.props.component;
+
     if (component === 'button' && this.props.href) {
       // If `href` is provided and a custom component is not, we render `<a>` instead
       component = 'a';
@@ -72,11 +129,14 @@ export class Button extends React.PureComponent {
     return component;
   }
 
-  classNames() {
+  classNames(): string {
     const variationClass = this.props.variation && `ds-c-button--${this.props.variation}`;
+
     const disabledClass =
       this.props.disabled && this.componentType() !== 'button' && 'ds-c-button--disabled';
+
     const sizeClass = this.props.size && `ds-c-button--${this.props.size}`;
+
     const inverseClass = (this.props.inversed || this.props.inverse) && 'ds-c-button--inverse';
 
     return classNames(
@@ -89,20 +149,20 @@ export class Button extends React.PureComponent {
     );
   }
 
-  handleKeyPress(e) {
+  handleKeyPress(e: React.KeyboardEvent): void {
     // Trigger onClick on space key event for `<a>` elements
     if (e.key === ' ') {
       this.handleClick(e);
     }
   }
 
-  handleClick(e) {
+  handleClick(e: React.MouseEvent | React.KeyboardEvent): void {
     if (!this.props.disabled && this.props.onClick) {
       this.props.onClick(e);
     }
   }
 
-  render() {
+  public render(): React.ReactNode {
     const attrs = this.attrs();
     const ComponentType = this.componentType();
 
@@ -117,55 +177,3 @@ export class Button extends React.PureComponent {
     );
   }
 }
-
-Button.defaultProps = {
-  type: 'button',
-  component: 'button',
-};
-Button.propTypes = {
-  /**
-   * Label text or HTML
-   */
-  children: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
-  /**
-   * Additional classes to be added to the root button element.
-   * Useful for adding utility classes.
-   */
-  className: PropTypes.string,
-  /**
-   * When provided, this will render the passed in component. This is useful when
-   * integrating with React Router's `<Link>` or using your own custom component.
-   */
-  component: PropTypes.oneOfType([PropTypes.element, PropTypes.elementType, PropTypes.func]),
-  disabled: PropTypes.bool,
-  /**
-   * When provided the root component will render as an `<a>` element
-   * rather than `button`.
-   */
-  href: PropTypes.string,
-  /**
-   * Access a reference to the `button` or `a` element
-   */
-  inputRef: PropTypes.func,
-  /** @hide-prop [Deprecated] Use inversed instead */
-  inverse: PropTypes.bool,
-  /** Applies the inverse theme styling */
-  inversed: PropTypes.bool,
-  /**
-   * Returns the [`SyntheticEvent`](https://facebook.github.io/react/docs/events.html).
-   * Not called when the button is disabled.
-   */
-  onClick: PropTypes.func,
-  size: PropTypes.oneOf(['small', 'big']),
-  /**
-   * Button [`type`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-type) attribute
-   */
-  type: PropTypes.oneOf(['button', 'submit']),
-  /**
-   * A string corresponding to the button-component variation classes.
-   * The `'danger'` variation is deprecated and will be removed in a future release.
-   */
-  variation: PropTypes.oneOf(['primary', 'danger', 'success', 'transparent']),
-};
-
-export default Button;
